Added support for Collecting metrics of EC2 hosts without the limitation of having i-012345678 in HOST.HOST
